fix minor problems
