--- src/interfaces/IUsers.ts
@@ -1,12 +1,0 @@
-export interface IUserItem {
-    id: number,
-    createdAt: string,
-    updatedAt: string,
-    email: string,
-    password: string,
-    role: string,
-}
-
-export interface IUsersData {
-    list?: Array<IUserItem>
-}
