--- src/interfaces/user-item.tsx
@@ -1,10 +1,0 @@
-export interface IUserItem {
-    id: number,
-    createdAt: string,
-    updatedAt: string,
-    email: string,
-    password: string,
-    role: string,
-    users: []
- }
-
