+++ src/pages/Users/Users.tsx
@@ -3,7 +3,7 @@ import UsersContainer from '../../components/Containers/Users-container';
 
 const Users: React.FC = () => {
    return (
-     <div><UsersContainer /></div>
+      <UsersContainer />
    )
 }
